Fix api database
